+++ examples/next-openai/app/api/chat-openai-fetch-pdf-custom-tool/route.ts
@@ -1,12 +1,13 @@
 import { openaiFetchPdfCustomToolAgent } from '@/agent/openai-fetch-pdf-custom-tool-agent';
-import { validateUIMessages } from 'ai';
+import { createAgentUIStreamResponse } from 'ai';
 
 export async function POST(request: Request) {
-  const body = await request.json();
+  const { messages } = await request.json();
 
-  console.dir(body.messages, { depth: Infinity });
+  console.dir(messages, { depth: Infinity });
 
-  return openaiFetchPdfCustomToolAgent.respond({
-    messages: await validateUIMessages({ messages: body.messages }),
+  return createAgentUIStreamResponse({
+    agent: openaiFetchPdfCustomToolAgent,
+    messages,
   });
 }