+++ packages/ai/src/agent/index.ts
@@ -1,22 +1,22 @@
 export { type Agent } from './agent';
-export { type BasicAgentOnFinishCallback } from './basic-agent-on-finish-callback';
-export { type BasicAgentOnStepFinishCallback } from './basic-agent-on-step-finish-callback';
+export { type ToolLoopAgentOnFinishCallback } from './tool-loop-agent-on-finish-callback';
+export { type ToolLoopAgentOnStepFinishCallback } from './tool-loop-agent-on-step-finish-callback';
 export {
-  type BasicAgentSettings,
+  type ToolLoopAgentSettings,
 
   /**
-   * @deprecated Use `BasicAgentSettings` instead.
+   * @deprecated Use `ToolLoopAgentSettings` instead.
    */
-  type BasicAgentSettings as Experimental_AgentSettings,
-} from './basic-agent-settings';
+  type ToolLoopAgentSettings as Experimental_AgentSettings,
+} from './tool-loop-agent-settings';
 export {
-  BasicAgent,
+  ToolLoopAgent,
 
   /**
-   * @deprecated Use `BasicAgent` instead.
+   * @deprecated Use `ToolLoopAgent` instead.
    */
-  BasicAgent as Experimental_Agent,
-} from './basic-agent';
+  ToolLoopAgent as Experimental_Agent,
+} from './tool-loop-agent';
 export {
   /**
    * @deprecated Use `InferAgentUIMessage` instead.
@@ -24,3 +24,6 @@ export {
   type InferAgentUIMessage as Experimental_InferAgentUIMessage,
   type InferAgentUIMessage,
 } from './infer-agent-ui-message';
+export { createAgentUIStreamResponse } from './create-agent-ui-stream-response';
+export { createAgentUIStream } from './create-agent-ui-stream';
+export { pipeAgentUIStreamToResponse } from './pipe-agent-ui-stream-to-response';